Terminando endpoints posts y comments

+++ back/src/modules/posts/service/posts.service.ts
@@ -1,19 +1,36 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Post } from '../entities/post.entity';
 import { Repository } from 'typeorm';
 import { CreatePostDto } from '../dtos/create-post.dto';
 import { UpdatePostDto } from '../dtos/update-post.dto';
+import { User } from 'src/modules/users/entities/user.entity';
 
 @Injectable()
 export class PostsService {
   constructor(
     @InjectRepository(Post)
     private readonly postRepository: Repository<Post>,
+    @InjectRepository(User)
+    private readonly userRepository: Repository<User>
   ) {}
 
-  create(createPostDto: CreatePostDto) {
-    const post = this.postRepository.create(createPostDto);
+  async create(createPostDto: CreatePostDto) {
+    //se verifica la existencia del user
+    const user = await this.userRepository.findOne({
+      where: { id: createPostDto.userId },
+    });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    //crear el post con el user asociado
+    const post = this.postRepository.create({
+      description: createPostDto.description,
+      image: createPostDto.image,
+      user: user, //asignamos el user al post
+    });
+    //guardamos el post en DB
     return this.postRepository.save(post);
   }
 
@@ -21,15 +38,45 @@ export class PostsService {
     return this.postRepository.find();
   }
 
-  findOne(id: string) {
-    return this.postRepository.findOne({ where: { id } });
+  async findOne(id: string): Promise<Post> {
+    const post = await this.postRepository.findOne({ 
+      where: { id }, 
+      relations: ['comments'],
+    });
+
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+
+    return post;
   }
 
-  update(id: string, updatePostDto: UpdatePostDto) {
-    return this.postRepository.update(id, updatePostDto);
+  async update(id: string, updatePostDto: UpdatePostDto) {
+    // Preload carga el post con el ID proporcionado y actualiza las propiedades
+    const post = await this.postRepository.preload({
+      id: id, 
+      ...updatePostDto,//Asignamos lso valores del dto al post
+    });
+    
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+    return this.postRepository.save(post);
   }
 
-  remove(id: string) {
-    return this.postRepository.delete(id);
+  async remove(id: string) {
+    const post = await this.postRepository.findOne({ where: { id } });
+
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+
+    const result = await this.postRepository.delete(id);
+
+    if (result.affected === 0) {
+      throw new InternalServerErrorException('Failed to delete post');
+    }
+
+    return { message: 'Post successfully deleted' };
   }
 }
